+++ run/beforeStart/10_configure.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 set -a
+: ${ENVIRONMENT:=production}
 : ${NODE_NAME:="VerneMQ@127.0.0.1"}
 : ${AUTH_ANONYMOUS:=off}
 : ${LOG_LEVEL:=info}
@@ -10,7 +11,7 @@ set -a
 : ${ADAPTER_HOST:=broker-adapter}
 : ${ADAPTER_PORT:=8080}
 
-cat /etc/vernemq/vernemq.conf.tpl | \
+cat /opt/broker/vernemq.conf.tpl | \
   envsubst '
     $NODE_NAME
     $AUTH_ANONYMOUS
@@ -22,4 +23,11 @@ cat /etc/vernemq/vernemq.conf.tpl | \
     $ADAPTER_HOST
     $ADAPTER_PORT
   ' > /etc/vernemq/vernemq.conf
-cat /etc/default/vernemq.tpl | envsubst > /etc/default/vernemq
+
+cat /opt/broker/vernemq.default.tpl | \
+  envsubst '
+    $ENVIRONMENT
+    $VMQ_VERSION
+    $ADAPTER_HOST
+    $ADAPTER_PORT
+  ' > /etc/default/vernemq