+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Generation of vernemq process environment variables
 
 
 [Unreleased]: https://gitlab.com/gfcc/dispatcher/compare/0.3.0...develop